BL-15 likes/comments/naviagtion for authorize user

+++ app/Http/Controllers/CommentController.php
@@ -6,6 +6,8 @@ use App\Models\Comment;
 use App\Models\Post;
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Auth;
+
 class CommentController extends Controller
 {
     public function index(Post $post)
@@ -26,18 +28,31 @@ class CommentController extends Controller
 
         $comment = new Comment([
             'content' => $request->input('content'),
-            'post_id' => $post->post_id,
-            'user_id' => session('current_user_id'), // ID користувача
+            'post_id' => $post->post_id,  
+            'user_id' => Auth::id(),
         ]);
+
         $comment->save();
 
-        return redirect()->route('comments.index', $post)->with('success', 'Коментар успішно додано.');
+        // Перевірка ролі користувача
+        if (Auth::user()->role === 'admin') {
+            return redirect()->route('comments.index', $post)->with('success', 'Коментар успішно додано.');
+        } else {
+            return redirect()->route('userPosts.show', $post)->with('success', 'Коментар успішно додано!');
+        }
     }
 
     public function destroy(Comment $comment)
     {
-        $comment->delete();
-        return redirect()->back()->with('success', 'Коментар успішно видалено.');
+        if (Auth::id() === $comment->user_id || Auth::user()->role === 'admin') {
+            $post = $comment->post;
+
+            $comment->delete();
+            return redirect()->route('userPosts.show', $post)->with('success', 'Коментар успішно видалено!');
+        }
+
+        return redirect()->back()->with('error', 'Ви не маєте права видаляти цей коментар.');
     }
+
 }
